last version of project(ready for deploy)

+++ houses/serializers.py
@@ -31,11 +31,18 @@ class ImageSerializer(serializers.ModelSerializer):
         fields = ('id', 'photo',)
 
 
+class DeveloperSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = UserProfile
+        fields = ('first_name', 'last_name', 'telephone', 'email', 'avatar')
+
+
 class HouseSerializer(serializers.ModelSerializer):
     advantage = AdvantagesSerializer(read_only=True)
     news = NewsSerializer(many=True, read_only=True, allow_null=True)
     documents = DocumentSerializer(many=True, read_only=True, allow_null=True)
     images = ImageSerializer(many=True, read_only=True, allow_null=True)
+    user = DeveloperSerializer(read_only=True)
 
     class Meta:
         model = House
@@ -44,7 +51,7 @@ class HouseSerializer(serializers.ModelSerializer):
             'house_type', 'house_class', 'building_technique', 'territory', 'distance_to_the_sea', 'communal_payments',
             'ceiling_height', 'gas', 'heating', 'type_of_sewerage', 'water_supply', 'agreements', 'payment_option',
             'house_purpose', 'amount_in_contract', 'department_first_name', 'department_last_name',
-            'department_telephone', 'department_email', 'advantage', 'news', 'documents', 'images'
+            'department_telephone', 'department_email', 'user', 'advantage', 'news', 'documents', 'images'
         )
 
 
@@ -107,9 +114,10 @@ class HouseImageSerializers(serializers.ModelSerializer):
         fields = ('image', 'delete_list')
 
     def update(self, instance, validated_data):
-        delete_list = validated_data.get('delete_list', False)
+
+        delete_list = validated_data.pop('delete_list', False)
         images_data = validated_data.pop('image', False)
-        if delete_list:
+        if delete_list and delete_list != ['']:
             Image.objects.filter(id__in=delete_list[0].split(',')).delete()
         if images_data:
             for image_data in images_data:
@@ -172,7 +180,3 @@ class ApartmentSerializer(serializers.ModelSerializer):
         return attrs
 # endregion Requests to add in house
 
-
-# region Chessboard
-
-# endregion Chessboard
